decrease method count
